+++ src/model/application/component.rs
@@ -253,10 +253,10 @@ pub struct InputText {
     pub label: Option<String>,
     /// Minimum input length for a text input; min 0, max 4000
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub min_length: Option<u64>,
+    pub min_length: Option<u16>,
     /// Maximum input length for a text input; min 1, max 4000
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_length: Option<u64>,
+    pub max_length: Option<u16>,
     /// Whether this component is required to be filled (defaults to true)
     #[serde(default = "default_true")]
     pub required: bool,
